# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vol d'Histoire API
  description: |
    API REST de l'encyclopédie de l'aviation militaire **Vol d'Histoire**.

    ## Authentification
    JWT en header `Authorization: Bearer <access_token>` (15 min) +
    refresh token en cookie HttpOnly (7 jours, rotation au refresh).

    ## Internationalisation
    Détectée via `?lang=fr|en` ou `Accept-Language`. Les fiches d'avions
    exposent `name_fr` / `country_name_fr` / `type_name_fr` en plus des
    champs traduits, comme **clés stables côté client** (utilisées dans les
    URLs de filtres pour rester valides indépendamment de la langue d'affichage).

    ## Anti-spam
    Les routes `/register` et `/auth/forgot-password` acceptent un token
    hCaptcha dans le body (`h-captcha-response`). Si `HCAPTCHA_SECRET`
    n'est pas configuré côté serveur, la vérification est skippée (mode dev).

    ## Cache & versioning
    Toutes les pages HTML sont servies avec un query string `?v=<version>`
    sur les assets pour bust le cache navigateur à chaque déploiement.
    Les statiques (CSS/JS/fonts) ont un `Cache-Control` long.
  version: 4.0.0
  contact:
    name: Vol d'Histoire
    url: https://vol-histoire.titouan-borde.com
servers:
  - url: /api
    description: API principale
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT (15 min). Obtenu via POST /login ou POST /refresh.
...
paths:
...
  /register:
    post:
      tags: [Auth]
      summary: Inscription
      description: Crée un compte et envoie un email de vérification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 255
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: "1 majuscule, 1 minuscule, 1 chiffre minimum"
                'h-captcha-response':
                  type: string
                  description: |
                    Token hCaptcha (optionnel). Vérifié côté serveur si
                    `HCAPTCHA_SECRET` est configuré, sinon ignoré.
      responses:
        '201':
          description: Compte créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  warning:
                    type: string
                    enum: [EMAIL_NOT_SENT]
                    description: Présent si l'email de vérification n'a pas pu être envoyé
        '400':
          description: "Données invalides, email déjà utilisé, ou captcha invalide (code: CAPTCHA_FAILED)"
        '429':
          description: Rate limit (10 req / 15 min)
  /login:
    post:
      tags: [Auth]
      summary: Connexion
      description: |
        Retourne un access token JWT (15 min) dans le body et un refresh token
        en cookie HttpOnly SameSite=Strict (7 jours).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Connexion réussie
          headers:
            Set-Cookie:
              description: refreshToken (HttpOnly, SameSite=Strict, Path=/api)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                    description: Access token JWT
        '400':
          description: Email ou mot de passe incorrect
        '403':
          description: Email non vérifié (code EMAIL_NOT_VERIFIED)
        '429':
          description: Rate limit (20 req / 15 min)
  /refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir le token
      description: |
        Émet un nouveau access token + rotation du refresh token.
        Le refresh token est lu depuis le cookie `refreshToken`.
      responses:
        '200':
          description: Nouveau token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Refresh token absent, expiré ou révoqué
...
  /auth/verify-email:
    get:
      tags: [Auth]
      summary: Vérifier l'email
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            pattern: '^[a-f0-9]{64}$'
      responses:
        '200':
          description: Email vérifié avec succès
        '400':
          description: Token invalide ou expiré
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Renvoyer l'email de vérification
      description: Réponse toujours 200 (pas de fuite d'information).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email envoyé (si le compte existe et n'est pas vérifié)
        '429':
          description: Rate limit (3 req / 15 min)
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Mot de passe oublié
      description: |
        Réponse toujours 200 (pas de fuite d'information sur l'existence du compte).
        Protégé par hCaptcha si `HCAPTCHA_SECRET` est configuré côté serveur.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                'h-captcha-response':
                  type: string
                  description: Token hCaptcha (vérifié si configuré).
      responses:
        '200':
          description: Email de réinitialisation envoyé (si le compte existe)
        '400':
          description: "Captcha invalide (code: CAPTCHA_FAILED) ou email malformé"
        '429':
          description: Rate limit (3 req / 15 min)
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Réinitialiser le mot de passe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                  pattern: '^[a-f0-9]{64}$'
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '200':
          description: Mot de passe réinitialisé
        '400':
          description: Token invalide ou mot de passe non conforme
        '429':
          description: Rate limit (5 req / 15 min)

  # ═══════════════════════════════════════════════════════════
  # USERS
  # ═══════════════════════════════════════════════════════════
...